Added Camera
This is the camera we will use for our app (MQTT).
