--- SmartcarMqttController/app/src/main/res/layout/activity_main.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <ImageView
-            android:id="@+id/imageView"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="0.3"
-            android:contentDescription="cameraView" />
-
-        <Button
-            android:id="@+id/button11"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="0.3"
-            android:textColor="#FAFAFA"
-            android:backgroundTint="#616161"
-            android:onClick="moveForward"
-            android:text="Forward" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="0.3"
-            android:orientation="horizontal">
-
-            <Button
-                android:id="@+id/button15"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:layout_weight="1"
-                android:textColor="#FAFAFA"
-                android:backgroundTint="#616161"
-                android:onClick="moveForwardLeft"
-                android:text="Left" />
-
-            <Button
-                android:id="@+id/button14"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:textColor="#FAFAFA"
-                android:backgroundTint="#212121"
-                android:onClick="stop"
-                android:text="Stop" />
-
-            <Button
-                android:id="@+id/button13"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:layout_weight="1"
-                android:textColor="#FAFAFA"
-                android:backgroundTint="#616161"
-                android:onClick="moveForwardRight"
-                android:text="Right" />
-        </LinearLayout>
-
-        <Button
-            android:id="@+id/button12"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="0.3"
-            android:textColor="#FAFAFA"
-            android:backgroundTint="#616161"
-            android:onClick="moveBackward"
-            android:text="Backward" />
-    </LinearLayout>
-</androidx.constraintlayout.widget.ConstraintLayout>
