--- SmartcarMqttController/settings.gradle
@@ -1,2 +1,0 @@
-include ':app'
-rootProject.name = "SmartcarMqttController"
